requirements.txt
streamlit   # for frontend
google-generativeai
python-dotenv
pathlib 
youtube_transcript_api
